+++ packages/visual-grid-client/test/e2e/openEyes.e2e.test.js
@@ -103,7 +103,10 @@ describe('openEyes', () => {
       scriptHooks,
     })
 
-    const results = await close()
+    const [errArr, results] = await presult(close())
+    errArr && console.log(errArr)
+    expect(errArr).to.be.undefined
+
     expect(results.length).to.eq(3)
     expect(results.map(r => r.getStatus())).to.eql(['Passed', 'Passed', 'Passed'])
   })